Patch 0.5
"We've polished 90% of the login page, we couldn't get some of the features working, so we decided to work on that another time because it would drag time if we ponder about it too much. But here are the rest of the details below"

- Kind of added a gradient image? (Which might be taken out the next patch since it looks weird.)
- Added a header image, it still needs to be fixed, which will happen at the next patch.

P.S. Some struggles happened today so we will take it easy for this patch, and hopefully we can do better at the next patch.

What's next?
- Complete the main page, or it doesn't have to be complete but we need to show good progress on it.