Patch 0.9 (Major changes)
"For the past few weeks, we have created some minor commits one at a time. But with the major holdup, we were able to make phenomenal progress on our website. There is still work to be done, but so far we are coming closer to that end goal."

- Developed the Home Page (95%) (Only needs some polishing now).
- Developed the Shop Page (50%)
- Product Page (50%)
- Developed the About Us Page (40%)
- Imported Product Images
- Imported our profile pictures for the about us page.

What's Next?
- Doing the front end of our website is something that can be completed much sooner than doing the back end. So me and my team are looking to focus on getting a better comprehension of doing the back end aspect of our website.